locks regular users out of pages that should require admin/coach permissions

+++ app/routes.py
@@ -126,6 +126,7 @@ def reset_password(token):
 
 
 @app.route("/tournament_creation", methods=["GET", "POST"])
+@login_required
 def tournament_creation():
     """
     If the league box is blank, we will take whichever league was selected from the dropdown for the
@@ -134,6 +135,9 @@ def tournament_creation():
     off all of the information put in.
 
     """
+    # if user not coach or admin, deny access by redirect
+    if not (current_user.is_coach or current_user.is_admin):
+        return redirect(url_for("access_denied"))
     leagues = League.query.all()
     form = TournamentCreationForm()
     if form.validate_on_submit():
@@ -189,6 +193,7 @@ def tournament_creation():
 
 
 @app.route("/tournament_dashboard", methods=["GET", "POST"])
+@login_required
 def tournament_dashboard():
     # TODO: rework the decorator style again by checking all filters and doing one big query instead of trying to make many small ones
     form = Search()
@@ -214,6 +219,7 @@ def tournament_dashboard():
     
 
 @app.route("/tournament_page", methods=["GET", "POST"])
+@login_required
 def tournament_page():
     tournament_string = request.args.get("tournament", None)
     tournament = Tournament.query.filter_by(tournament_name=tournament_string).first()
@@ -241,6 +247,9 @@ def tournament_page():
         # Delete button will delete the tournament from the database and then return the the tournament dahsboard.
         # This will also clear the relations from the teams table.
         elif request.form.get("delete_button") == "Delete Tournament":
+            # if user not coach or admin, deny access by redirect
+            if not (current_user.is_coach or current_user.is_admin):
+                return redirect(url_for("access_denied"))
             tournament.tournament_teams.clear()
             db.session.delete(tournament)
             db.session.commit()
@@ -250,6 +259,9 @@ def tournament_page():
             )
             return redirect(url_for("tournament_dashboard"))
         elif request.form.get("register_button") == "Register":
+            # if user not coach or admin, deny access by redirect
+            if not (current_user.is_coach or current_user.is_admin):
+                return redirect(url_for("access_denied"))
             # Register will take the team the coach has with the same league, and register that team inside of the tournament.
             for team in all_teams:
                 if (
@@ -281,7 +293,11 @@ def tournament_page():
 
 
 @app.route("/tournament_management", methods=["GET", "POST"])
+@login_required
 def tournament_management():
+    # if user not coach or admin, deny access by redirect
+    if not (current_user.is_coach or current_user.is_admin):
+        return redirect(url_for("access_denied"))
     # Tournament management essentially does the same thing as the create team, but instead of making new tournament objects,
     # we are just adding on to the current tournament. With the way leagues are right now, I think we might have to remove
     # the option to edit leagues once a tournament is created for simplicity sakes.
@@ -343,11 +359,19 @@ def team(team_ID: int):
 
 
 @app.route("/<match_ID>/match", methods=["GET", "POST"])
+@login_required
 def match(match_ID: int):
     return render_template("match.html")
 
 
+@app.route("/access_denied")
+@login_required
+def access_denied():
+    return render_template("access_denied.html")
+
+
 @app.route("/league", methods=["GET", "POST"])
+@login_required
 def league():
     # if the user doesn't have an affiliated team and league,
     # the template will display a form so they can select one.
@@ -376,6 +400,9 @@ def league():
 @app.route("/create_team", methods=["GET", "POST"])
 @login_required  # TODO: It would be nice to have coach_required and admin_required decorators for these pages.
 def create_team():
+    # if user not coach or admin, deny access by redirect
+    if not (current_user.is_coach or current_user.is_admin):
+        return redirect(url_for("access_denied"))
     # TODO: Might want to update this later when coach and admin classes are
     # defined/we have a coaches table.
     coaches = User.query.filter_by(is_coach=True)
@@ -401,76 +428,89 @@ def create_team():
 
 
 @app.route("/manual_permissions", methods=["GET", "POST"])
+@login_required
 def manual_permissions():
+    # if user not coach or admin, deny access by redirect
+    if not (current_user.is_coach or current_user.is_admin):
+        return redirect(url_for("access_denied"))
+
     form = ManualPermissionsForm()
     users = db.session.query(User).order_by("id")
     prs = db.session.query(PermissionRequest).order_by("id")
 
-    tval = prs
     if form.validate_on_submit():
-        # user = User.query.filter_by(id=form.userID.data).first()
         pr = PermissionRequest.query.filter_by(id=form.prID.data).first()
         user = User.query.filter_by(id=pr.user).first()
-        tval = pr.id
-
-        """
-        if form.actions.data == 1:
-            approve_coach(current_user, user)
-        if form.actions.data == 2:
-            deny_coach(current_user, user)
-        if form.actions.data == 3:
-            approve_admin(current_user, user)
-        if form.actions.data == 4:
-            deny_admin(current_user, user)
-        """
-
-        if form.pr_actions.data == 1:
+        if form.pr_actions.data == 1:  # approve
             if pr.coach_request == 1:
                 approve_coach(current_user, user, pr)
             elif pr.admin_request == 1:
                 approve_admin(current_user, user, pr)
-        if form.pr_actions.data == 2:
+        if form.pr_actions.data == 2:  # deny
             if pr.coach_request == 1:
                 deny_coach(current_user, user, pr)
             elif pr.admin_request == 1:
-                deny_admin(current_user, user, pr)
-
-    return render_template(
-        "manual_permissions.html",
-        title="Permissions",
-        form=form,
-        users=users,
-        prs=prs,
-        tval=tval,
-    )
+                if current_user.id == user.id:
+                    flash("Do not try to remove your own admin permission!")
+                else:
+                    deny_admin(current_user, user, pr)
+    return render_template("manual_permissions.html", title="Permissions", form=form, users=users, prs=prs)
 
 
 @app.route("/request_permission", methods=["GET", "POST"])
+@login_required
 def request_permission():
     form = RequestPermissionsForm()
-    prs = PermissionRequest.query.filter_by(user=current_user.id).all()
     if form.validate_on_submit():
-        if form.actions.data == 1:
-            # generate new pr object
-            pr = PermissionRequest(coach_request=1, user=current_user.id)
-            db.session.add(pr)
-            db.session.commit()
-        if form.actions.data == 2:
-            pr = PermissionRequest(admin_request=1, user=current_user.id)
-            db.session.add(pr)
-            db.session.commit()
-    return render_template(
-        "request_permission.html", title="Request Permission", form=form, prs=prs
-    )
+        # prs = list of permission requests associated with current user
+        prs = PermissionRequest.query.filter_by(user=current_user.id).all()
+        coach_request = 0  # 1 if request already exists
+        admin_request = 0
+        for each in prs:
+            if each.label == 'Coach request':
+                coach_request = 1
+            if each.label == 'Admin request':
+                admin_request = 1
+        if form.request_coach.data:
+            if current_user.is_coach:
+                flash("You are already a Coach!")
+            elif coach_request == 0:
+                # generate new pr object
+                pr = PermissionRequest(coach_request=1,
+                                       user=current_user.id,
+                                       username=current_user.username,
+                                       label='Coach request')
+                db.session.add(pr)
+                db.session.commit()
+            else:
+                flash("You already have a Coach request in!")
+        if form.request_admin.data:
+            if current_user.is_admin:
+                flash("You are already an Admin!")
+            elif admin_request == 0:
+                pr = PermissionRequest(admin_request=1,
+                                       user=current_user.id,
+                                       username=current_user.username,
+                                       label='Admin request')
+                db.session.add(pr)
+                db.session.commit()
+            else:
+                flash("You already have an Admin request in!")
+    prs = PermissionRequest.query.filter_by(user=current_user.id).all()
+    return render_template("request_permission.html", title="Request Permission", form=form, prs=prs)
 
 
 @app.route("/dbtest", methods=["GET", "POST"])
+@login_required
 def dbtest():
+    # if user not coach or admin, deny access by redirect
+    if not (current_user.is_coach or current_user.is_admin):
+        return redirect(url_for("access_denied"))
     form = dbtestForm()
-    users = db.session.query(User).order_by("id")
-    teams = db.session.query(Team).order_by("id")
-    leagues = db.session.query(League).order_by("id")
-    tournaments = db.session.query(Tournament).order_by("id")
+    users = db.session.query(User).order_by('id')
+    teams = db.session.query(Team).order_by('id')
+    leagues = db.session.query(League).order_by('id')
+    tournaments = db.session.query(Tournament).order_by('tournament_id')
     # test value
     tval = "none"
     models = {
